mcp_server/requirements.txt
aiohttp
asyncio
certifi
mcp
mcp[cli]
fastapi
uvicorn
python-dotenv